Added interfaces for eBPF instructions
Added the Jumper and Valuer interfaces which allows someone to check if
an instruction is a branching instruction without having to look at the
op code or checking all Jump instruction types.

By defining setters as part of the interface definition we also allow
setting of the most important fields which makes implementing linkers
much easier.