Made ebpf package more uniform
Removed 'If' from instruction struct names so all Jump instructions are
in the form Jump{operand}. Also renamed Val fields to Value, so all
structs now use the same field name.

Also changed all instructions functions to pointer receivers since the
Jumper interface may need to change an instruction. If this happens on a
non-pointer implementation of Instruction, this causes issues. So from
now on, only the pointer variant of the struct implements the
Instruction interface.

Lastly, added byte order functions. Its technically outside of the scope
of eBPF itself, but they are frequently required when generating eBPF
programs using the ebpf package. So it makes sense to just have the
functions in there.