Added assembly parsing to ebpf package
The main purpose of the ebpf package is to enable dynamic eBPF program
generation. Using the instruction structs directly allows for
precise generation, setting the exact values as machine code. But it is
very verbose and thus difficult to work with.

Generating assembly code and turning that into eBPF instructions has
some advantages over direct instruction generation. The assembly is
easier to read than an array of go structs and the usage of labels
eliminates the need to change offset values in jumps when code is
reordered.

Currently the assembler only produces instruction, not yet full
programs/ELF files. So the locations of Map FD's must still be tracked
externally until the assembler can emit relocation tables.