Added tracepoint program support
Added the ability to attach BPF programs to tracepoints. This is the
first step, next up is kprobes and uprobes. The perf package was added
to facilitate all logic surrounding the perf subsystem. The standard
library doesn't provide wrappers for access to perf_event_open syscall
nor does the x/sys/unix library. Google didn't yeld any Golang libraries
for the perf subsystem which can also be used without BPF, so I might
spin the perf package of into its own repo at some point.

I am about to start a large refactor of the BPFProgram type and change
it to be similair to the BPFMap where we have an generic interface,
abstract type and specific subtypes. So commiting this first.