Refactored BPFPrograms, introduced specific progam types
In order to better support different program types and attach methods I
have refactored the way BPF programs are represented. All common
functionality lives in the abstract program which is embedded in all
specific programs. Every program type has its own type.

When loading from ELF we use the section names as hints for the program type, if this
is incorrect for some reason it is up to the loader program(user of this
library) to change the type.

Users will need to cast the BPF Program before they can use program type
specific features. This shouldn't be to much of a hassle since most
users will know what kind of program type they will be loading.