Added a map in map example and made array and hash of maps work
Created a map in map example. Turns out the existing map in map types
didn't work since map-in-map maps require the innerMap field to be set
during map loading.

Also added a global map register, which holds references to all existing
eBPF maps. This was needed since the map-in-map maps return IDs to the
maps instead of FD's like you would think they do. If you use the
ID-to-FD syscall it creates a new FD. This is undesirable because this
new instance needs to be closed seperatly from the original. By
maintaining a global register with ID's and FD's we can always return
the same BPFMap instance as the original with the same FD. This is more
logicical for users of the API and hopefully will avoids FD leaks.