Added BTF support for programs
This is the first iteration, some work still needs to be done on the BTF
feature. Parsing of .BTF, and .BTF.ext ELF sections works. Relocations
in BTF are handled stragely, the offsets didn't seem to make sense, so
left that out for now. Seems like it doesn't matter for the current
examples but is required for CO:RE.

Added a new BTF struct, which like the programs and maps can be loaded
into the kernel. It still requires a bit of cleanup. Currently
generating BTF from go is not yet supported. At some point I want to be
able to marshal the high level structs back to byte slices, but for now
BTF can only be used when loaded from ELF files.

BTF loading for programs is fully working. Maps, however, don't work.
They are still a challange. It looks like users need to use the libbpf
helper macros to get clang to include type information in map definitions.