Added first iteration of testing suite
Added the first iteration of the BPF testing suite. This version still
doesn't work for everyone since it relies on local files. This will be
fixed in a later iteration.

We want to test the library on multiple kernel versions and on different
CPU architectures. We also might want to test some BPF features which
might disrupt tasks of the OS(blocking network packets). Because of that
testing is preformed in a VM using QEMU, which should allow us to
rapidly test multiple kernels and CPU architectures(via emulation).

The test suite relies on pre-built linux kernels and an
initrd archive. Build scripts for these and the pre-built results will
be placed in a seperate repo. We will start making use of these publicly
hosted pre-built images in the second iteration.

The test suite works by generating binaries on the host using `go test`
and then running them in a QEMU VM. Currently this will only work on
linux since creating the QEMU disk image requires some linux foo(parted, mkfs, loop devices). Maybe
some day we will implement this is test suite directly. Then a VM is
started with a disk attached containing the test binaries and a script
which runs them with the correct flags. Any results like the stdout,
profiles, traces, coverage will be writen to the mounted disk. After all
tests are done the VM will exit. The files can be extracted from the
disk and processed by the testsuite on the host machine.

The result should be a test tool which can run
compatibility/unit/intergation tests with just one command.