Testsuite now returns useful output and exit code
The test suite now return the results of each individual test. If any of
the environments fail a non-0 exit code is returned, CI's can check the
exit code to know if anything went wrong.

Added the --failfast and --short flags, which do the same as the `go
test` versions. The only difference is that --failfast also applies to
the environments.

Lastly made some changes to the unit tests so they now skip if the
features under test are not supported by the kernel