Added kernel 4.14.260 to test suite
After adding this kernel to the test suite, almost everything broke :)
Turns out that the kernel doesn't gracefully handle oversized attributes
to the bpf syscall, at least not in older versions. So had to implement
version based size calculation for the program load sub command.

Exported the current kernel version directly, instread of just the
infered kernel features. Did this since, in cases like the attr size
check, it makes more sense to compare against the kernel version
directly.

Also discovered that until version 4.16, the get next key for LPM maps
is not supported, so added it to kernelversion so tests that rely on
iteration can skip it. Also added an errIterator, which can be returned
in case map iteration is not supported at all.