Added global data relocation support
We now support global data loading from .data/.rodata/.bss ELF sections.
Values declared in global scope are put in .data of they are writable,
in .rodata if they are read only(const) and .bss if they are writable
but not initialized to a specific value

Clang will not make relocation entries to symbols but to the whole
section. The offset into that section is stored in the ldimm64
instruction. The actual data contained in the ELF section is loaded as
an array map with a single key, its value the size of the data.

The ldimm64 instructions are loaded with both the map FD and the offset,
its source set to BPF_PSEUDO_MAP_FD_VALUE. This allows direct memory
access into the map.

To verify that all of this works as expected, the global data selftest
from the kernel was adopted and added as an intergration test. Since
this test uses strings, support for arrays and slice as value pointers
to map operations was added.