Added testsuite chmod'ing and sudo path lookup
The testsuite has been outputting results as root, it does this because
we always run the executable as root. But having the resulting files as
root is difficult to work with since we can't modify or delete the
files. So we now chmod all result files to the uid:gid with which the
program was originally called. We rely on the environment variables set
by sudo to do this.

Also added proper executable path lookup via the PATH environment
variable, since we can't rely on a hardcoded absolute path.