Added a number of additional emulator features
Added the following features to the emulator:
- Maps support
- Array map
- BPF-to-BPF calls
- Helper function
- bpf_map_lookup_element helper function
- A few instructions

The emulator seems to be sufficiantly functional to be merged into
master. The eBPF debugger which is the reason for having the emulator
will be living in a seperate repository, since it is really a tool on
its own, which ofcource uses this library.