Added BTF to program even if the ELF contains no relocations
In the ELF parser linkAndRelocate phase we set the BTF on a program
since we modify the BTF during relinking. But if a program doesn't need
to be relocated or linked the BTF object whould never be set.

We now always initially set the BTF object, and replace it if the
program is relocated and/or linked