Added BPF_MAP_TYPE_HASH and bpf_update_element emulator support
Added the hash, map type. It is not the most efficient implementation
but it does work. That is because we can't use byte slices directly in
go map's, to work around this we hash all slices which always results in
a fixes size array(32 bytes in the case of sha256 which was chosen).
This means that there is a double hashing cost and also that this map
takes up way more space then nessesery. Optimization is currently not
the focus, might in the future be done by using a custom hash map
instread of the built-in hash map.

Added the `bpf_update_element` map helper function, which is a light
wrapper around the map.Update function. Decided to remove linux style
error-number returns from the maps themselfs, instread we now return
pre-defined errors for specific situations which the helper functions
have to translate into linux specific error codes. Since someone using
this VM for non-linux things shouldn't have to deal with linux-like
error codes.