Added LRU map and "fake" per CPU maps
Added an emulated version of the BPF_MAP_TYPE_LRU_HASH map type. We use
a slice of key hashes to track recent usage of keys. Not very efficient,
a linked list of sorts would be more efficient, something to think about
later.

The AbstractMapToVM function now will return the normal varients of the
emulated maps when they are requested. Did this since there is no
paractical difference for eBPF programs as long as we do not yet support
concurrent program execution within the same VM.

As soon as we support concurrent execution, we should also make seperate
map types.