Added saving and restoring of callee registers(R6-R9)
It was my understanding that the eBPF calling convention is an actual
convention and that it was up to the compiler to not use R6 to R9 in
function calls. Turns out this is not true, the eBPF virtual machine is
suppored to save the R6-R9 registers and restore them upon return from a
function call, just like the PC register. Functions are free to do
whaterver they want with callee saved registers. So added this feature.