Fixed bug caused by non-commutative add operation
Discovered a bug because the add operation of two registers was not
commutative. When adding a plain value to a pointer, we would modify the
offset of the pointer, the result being a pointer. But if we add a
pointer to a numeric value, the number value changed but it would not be
come a pointer.

This commit fixes that so that if the SRC register of an add operation
is a pointer, the resulting DST register will also become a pointer.