Emulator jump instruction check type equality
This commit adds type checking to jump instructions. It is very common
for programs to check if the return value of a function is not NULL or
the value 0. If the return type is always a plain value, this is no
issue. But if a function would return a pointer with offset 0, jump
instruction compared it to a 0 numeric value and returned true.

This is because when evaluating pointer types as int64, we just return
the offset into a memory address, since this is the property which can
be modified. For example ptr+12 + 10 = ptr+22, this works by taking 12
and 10, then adding and assinging it to the offset of ptr. A quirk of
the way we emulate memory, since in a normal "computer" a pointer would
be the actual memory address of the thing we are pointing to.

The way I "fixed" this is to assert that only variables of the same
'type' can be compared in branching instructions. So comparing two
pointer values is valid and two plain values is valid, but comparing two
dissimilar types will always retult in 'false', except for the
jump-not-equal which will always jump since it is the only inverted
instruction.