Added "old school" BTF map k/v type annotation support
The "old school"(pre-2019) way to add Key/Value BTF information to maps
was to use a BPF_ANNOTATE_KV_PAIR macro
https://elixir.bootlin.com/linux/v5.16.5/source/tools/testing/selftests/bpf/bpf_legacy.h#L5
This macro would add additional types to the eBPF program with the
actual key and value types instread of just the size we get from the
"maps" ELF segment.

The new way of doing things would be to declare a struct in the ".maps"
ELF segment which doesn't conform to the `bpf_map_def` memory layout.
The loader is then supposed to reconstruct the `bpf_map_def` from the
BTF info of the map. Fields like `value` and `key` whould be used with
actual data types not just their sizes, since the key and value sizes
can be infered from the BTF. This way of doing things requires some
rework of the ELF parsing, so will save this for later. Both methods
should be supported anyway.