Made maps and program indexes start from 1 instread of 0
zero is the default value for array maps or imm values of un-relocated
map FD load instructions. Normally this is quickly detected because in
linux the 0'th file descriptor is never assigned since it is always in
use by already and in kernel space these numbers are pointers to the
actual maps. In both cases 0 is invalid, but in the emulator is was
valid, which makes it really unclear if something is expected or a bug
caused by having a 0 value in a map or program index.

So from now on all 0 ocurrences for map and program id's are bad and we
will error if we catch any of those conditions