Changed elf maps indexing to use full names
Until now we always used the truncated map names as keys in the maps map
which is returned when parsing an ELF file. There is no good reason for
doing so since we only really need to truncate the name before loading
into the kernel. So we now use the full names.

Currently the elf parser will still complain if the TruncateNames option
is not set, and the names in the maps themselfs are also still
truncated. In a future commit we should just make everything simpler by
always using strings of unlimited lengths and truncating them on load.