Added map-in-map support to the emulator
Again, we reuse the normal array and hash map types for the map-in-map
types. This works because we just write the map ids/fds into the generic
array and hash maps. The only changed needed to make this work was to
make map-related helper functions accept pointer types as well, where
before they always expected a numeric type.