Added queue and stack map support to the emulator
Added the queue and stack map types and the push, pop and peek helper
functions used to interact with this map type.

The implementation of the queue is still inefficient, we currently do a
lot of copying using a slice, this should be replaced with a ringbuffer
at some point.